requirements.txt
flask
python-dotenv
pydantic
flask-sqlalchemy
psycopg2-binary
email-validator
pyjwt
bcrypt